Create Sentiment
This VBA (Visual Basic for Applications) code defines a subroutine called "Sub Sentiment_analysis()" that performs sentiment analysis on a range of cells in an Excel worksheet.

The code begins by declaring and initializing variables, including a range object (oRng), an array (T1) containing positive sentiment words, an index variable (i), and a sentiment score variable (sentScore).

The range object (oRng) is set to refer to a range of cells in the active worksheet, starting from cell A2 and extending to the last non-empty cell in column A.

The code then loops through the positive sentiment words array (T1) and uses the Find method to search for each word within the specified range. If a match is found, the sentiment score (sentScore) is incremented by 1.

Next, the code reassigns the array (T1) to contain negative sentiment words and performs a similar search loop. If a negative sentiment word is found, the sentiment score (sentScore) is decremented by 1.

Finally, a message box is displayed showing the calculated sentiment score.

In summary, this code analyzes the sentiment of the text in a specified range of cells and calculates a sentiment score based on the occurrence of positive and negative sentiment words within that range.